task

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -10,7 +10,7 @@
 
 int main(void)
 {
-int prime;
+long long int prime;
 prime = 1231952;
 int i;
 i = 2;
